+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for MXFP8 Linear layers via torchao.
 - Added support for tracking total flops.
 - Added support for Gemma 3 models.
+- Added support for Muon and Dion optimizers.
+- Added support for Ulysses-style context parallelism.
+- Added 60M and 1M model sizes.
 
 ### Fixed
 
@@ -35,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix overflow when too many global flops are computed.
 - Ladder lmevaluator typo.
 - Made some functions involved in data loading preprocessing more robust to race conditions.
+- GAPMonitorCallback would raise an error if a local tensor shard had 0 elements.
 
 ### Changed